--- src/main/java/com/example/springbootboard/user/repository/JpaUserRepository.java
@@ -1,9 +1,0 @@
-package com.example.springbootboard.user.repository;
-
-import com.example.springbootboard.user.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface JpaUserRepository extends JpaRepository<User, Long> {
-}
